Alfaversioon
Parandasin kirjavead ja muud probleemid funktsioonides, mis takistasid korrektset tööd.
Taastasin sissekannete loetelu ja avamise funktsioonid, uute käskudega "?" ja "!".
Lisasin debugimise lihtsustamiseks faili debugfail.txt, millega kohe pärast kloonimist lihtsamini käske testida.
